requirements.txt
altair==5.5.0
altgraph==0.17.4
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
blinker==1.9.0
boto3==1.40.55
botocore==1.40.55
bottle==0.13.4
cachetools==6.2.0
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
clr_loader==0.2.7.post0
colorama==0.4.6
diff-match-patch==20241021
dj-database-url==3.0.1
Django==5.2.7
django-import-export==4.3.12
django-jazzmin==3.0.1
django-storages==1.14.2
django-widget-tweaks==1.5.0
gitdb==4.0.12
GitPython==3.1.45
google-auth==2.41.1
google-auth-oauthlib==1.2.2
gspread==6.2.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
narwhals==2.7.0
numpy==2.3.3
oauth2client==4.1.3
oauthlib==3.3.1
packaging==25.0
pandas==2.3.3
pefile==2023.2.7
pillow==11.3.0
protobuf==6.32.1
proxy_tools==0.1.0
psycopg2-binary==2.9.11
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydeck==0.9.1
pyinstaller==6.16.0
pyinstaller-hooks-contrib==2025.9
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.1
python-telegram-bot==22.4
pythonnet==3.0.5
pytz==2025.2
pywebview==6.0
pywin32-ctypes==0.2.3
referencing==0.36.2
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.27.1
rsa==4.9.1
s3transfer==0.14.0
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlparse==0.5.3
streamlit==1.50.0
streamlit-autorefresh==1.0.1
tablib==3.9.0
tenacity==9.1.2
toml==0.10.2
tornado==6.5.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
waitress==3.0.2
watchdog==6.0.0
whitenoise==6.11.0
PyGithub==2.4.0

